Search Fields Added
Search Fields added, still need styling
