+++ averagecollector/src/main/java/com/chrisrecinos/model/data/controller/StartController.java
@@ -1,7 +1,9 @@
 package com.chrisrecinos.model.data.controller;
 
+import com.chrisrecinos.model.data.entity.CardSet;
 import com.chrisrecinos.model.data.entity.Team;
 import com.chrisrecinos.model.data.repository.TeamRepository;
+import com.chrisrecinos.model.data.services.CardSetService;
 import com.chrisrecinos.model.data.services.TeamService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
@@ -9,7 +11,10 @@ import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestMethod;
 
+import java.util.ArrayList;
+import java.util.HashMap;
 import java.util.List;
+import java.util.Map;
 
 /**
  * @author - Christopher Recinos
@@ -22,6 +27,9 @@ public class StartController {
     @Autowired
     private TeamService teamService;
 
+    @Autowired
+    private CardSetService cardSetService;
+
     @RequestMapping(method = RequestMethod.GET)
     String getIntroPage() {
         return "index";
@@ -30,10 +38,17 @@ public class StartController {
     @RequestMapping(value = "/homepage", method = RequestMethod.GET)
     String navToHomePage(Model model) {
         model.addAttribute("teams", getAllTeams());
+        model.addAttribute("sets", getAllSets());
         return "homepage";
     }
 
     private List<Team> getAllTeams() {
         return this.teamService.getTeams("", "");
+    }
+
+    private List<CardSet> getAllSets() {
+        List<CardSet> sets = this.cardSetService.getSets(null, null, null);
+
+        return cardSetService.sortYears(sets);
     }
 }
