+++ averagecollector/src/main/resources/data.sql
@@ -72,7 +72,7 @@ INSERT INTO CARD (SET_ID, PLAYER_ID, TEAM_ID, CARD_NUM, INSERT_TYPE, MEM_TYPE) V
                                                                                       (16, (SELECT id FROM PLAYER WHERE FIRST_NAME = 'George' AND LAST_NAME = 'Springer'),
                                                                                           (SELECT id FROM TEAM WHERE TEAM_NAME = 'Astros'), 'GSP', 'Full-Size Relic Design B', 'Relic'),
                                                                                       (11, (SELECT id FROM PLAYER WHERE FIRST_NAME = 'Magglio'),
-                                                                                          (SELECT id FROM TEAM WHERE TEAM_NAME = 'White Sox'), '23', 'Hit Kings', 'Bat'),
+                                                                                          (SELECT id FROM TEAM WHERE TEAM_NAME = 'White Sox'), '23', 'Hit Kings', 'Relic'),
                                                                                       (12, (SELECT id FROM PLAYER WHERE FIRST_NAME = 'Todd' AND LAST_NAME = 'Walker'),
                                                                                           (SELECT id FROM TEAM WHERE TEAM_NAME = 'Rockies'), 'TOWA', 'National Patch Time', 'Relic'),
                                                                                       (13, (SELECT id FROM PLAYER WHERE FIRST_NAME = 'Randy' AND LAST_NAME = 'Johnson'),
@@ -533,7 +533,7 @@ INSERT INTO CARD (SET_ID, PLAYER_ID, TEAM_ID, CARD_NUM, INSERT_TYPE, SERIAL_NUM,
                                                                                                   (63, (SELECT id FROM PLAYER WHERE LAST_NAME = 'Perry'),
                                                                                                       (SELECT id FROM TEAM WHERE TEAM_NAME = 'Tigers'), '368', 'Rookie Autographs', '025/300', 'Autograph'),
                                                                                                   (67, (SELECT id FROM PLAYER WHERE LAST_NAME = 'Rupp'),
-                                                                                                      (SELECT id FROM TEAM WHERE TEAM_NAME = 'Phillies'), 'CR', 'Debut Cuts', '009/200', 'Cut Autograph'),
+                                                                                                      (SELECT id FROM TEAM WHERE TEAM_NAME = 'Phillies'), 'CR', 'Debut Cuts', '009/200', 'Cut Signature'),
                                                                                                   (67, (SELECT id FROM PLAYER WHERE FIRST_NAME = 'Yonder'),
                                                                                                       (SELECT id FROM TEAM WHERE TEAM_NAME = 'Reds'), 'YA', 'Futures Game Materials', '159/199', 'Relic'),
                                                                                                   (80, (SELECT id FROM PLAYER WHERE FIRST_NAME = 'Billy' AND LAST_NAME = 'Williams'),
